pyproject.toml
[tool.poetry]
name = "watermarker"
version = "0.1.0"
description = ""
authors = ["morveusz93 <morveusz93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^9.4.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
